requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
albumentations==1.3.0
alembic==1.14.1
altair==5.5.0
apturl==0.5.2
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
black==22.3.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2020.6.20
cfgv==3.4.0
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
colorama==0.4.4
command-not-found==0.3
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
databricks-cli==0.18.0
dbus-python==1.2.18
decorator==5.2.1
defer==1.0.6
dill==0.3.9
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==6.1.3
docker-pycreds==0.4.0
dotty-dict==1.3.1
entrypoints==0.4
exceptiongroup==1.2.2
executing==2.2.0
fastcore==1.7.29
filelock==3.17.0
fire==0.1.3
Flask==2.3.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
greenlet==3.1.1
grpcio==1.70.0
gunicorn==20.1.0
httplib2==0.20.2
huggingface-hub==0.29.1
identify==2.6.8
idna==3.3
imageio==2.37.0
imgaug==0.4.0
importlib-metadata==4.6.4
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.5
joblib==1.4.2
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.6
keyring==23.5.0
kiwisolver==1.4.8
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lightning-utilities==0.12.0
loguru==0.7.3
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.3.1
monai==1.0.1
more-itertools==8.10.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.28.0
netifaces==0.11.0
networkx==3.4.2
nodeenv==1.9.1
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.11.0.86
packaging==23.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.8.0
Pillow==9.1.1
platformdirs==4.3.6
pre-commit==3.1.1
promise==2.3
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==3.12.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.4
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pytorch-lightning==2.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==6.0
qudida==0.0.4
querystring-parser==1.2.4
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.2
scikit-image==0.22.0
scikit-learn==1.6.1
scipy==1.8.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentry-sdk==2.22.0
setproctitle==1.3.5
shapely==2.0.7
shortuuid==1.0.13
six==1.16.0
smmap==5.0.2
snntorch==0.9.4
SQLAlchemy==2.0.38
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
systemd-python==234
tabulate==0.9.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.5.0
tifffile==2025.2.18
tokenizers==0.21.0
tomli==2.2.1
torch==1.13.1
torch-snippets==0.496
torchmetrics==0.10.3
torchvision==0.14.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
virtualenv==20.29.2
wadllib==1.3.6
wandb==0.13.4
wasabi==1.1.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
yarl==1.18.3
zipp==1.0.0
